Move package of QueryHeader (#16703)
