+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/response/header/query/QueryHeaderBuilderEngine.java
@@ -27,7 +27,6 @@ import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.infra.executor.sql.execute.result.query.QueryResultMetaData;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.infra.rule.identifier.type.DataNodeContainedRule;
-import org.apache.shardingsphere.proxy.backend.response.header.query.impl.QueryHeader;
 
 import java.sql.SQLException;
 
@@ -50,9 +49,10 @@ public final class QueryHeaderBuilderEngine {
      * @throws SQLException SQL exception
      */
     public QueryHeader build(final QueryResultMetaData queryResultMetaData, final ShardingSphereMetaData metaData,
-                                   final int columnIndex, final LazyInitializer<DataNodeContainedRule> dataNodeContainedRule) throws SQLException {
-        QueryHeaderBuilder builder = QueryHeaderBuilderFactory.newInstance(databaseType);
-        return builder.build(queryResultMetaData, metaData, queryResultMetaData.getColumnName(columnIndex), queryResultMetaData.getColumnLabel(columnIndex), columnIndex, dataNodeContainedRule);
+                             final int columnIndex, final LazyInitializer<DataNodeContainedRule> dataNodeContainedRule) throws SQLException {
+        String columnName = queryResultMetaData.getColumnName(columnIndex);
+        String columnLabel = queryResultMetaData.getColumnLabel(columnIndex);
+        return QueryHeaderBuilderFactory.newInstance(databaseType).build(queryResultMetaData, metaData, columnName, columnLabel, columnIndex, dataNodeContainedRule);
     }
     
     /**
@@ -67,10 +67,10 @@ public final class QueryHeaderBuilderEngine {
      * @throws SQLException SQL exception
      */
     public QueryHeader build(final ProjectionsContext projectionsContext, final QueryResultMetaData queryResultMetaData,
-                                    final ShardingSphereMetaData metaData, final int columnIndex, final LazyInitializer<DataNodeContainedRule> dataNodeContainedRule) throws SQLException {
-        QueryHeaderBuilder builder = QueryHeaderBuilderFactory.newInstance(databaseType);
-        return builder.build(queryResultMetaData, metaData, getColumnName(projectionsContext, queryResultMetaData, columnIndex), 
-                getColumnLabel(projectionsContext, queryResultMetaData, columnIndex), columnIndex, dataNodeContainedRule);
+                             final ShardingSphereMetaData metaData, final int columnIndex, final LazyInitializer<DataNodeContainedRule> dataNodeContainedRule) throws SQLException {
+        String columnName = getColumnName(projectionsContext, queryResultMetaData, columnIndex);
+        String columnLabel = getColumnLabel(projectionsContext, queryResultMetaData, columnIndex);
+        return QueryHeaderBuilderFactory.newInstance(databaseType).build(queryResultMetaData, metaData, columnName, columnLabel, columnIndex, dataNodeContainedRule);
     }
     
     private String getColumnLabel(final ProjectionsContext projectionsContext, final QueryResultMetaData queryResultMetaData, final int columnIndex) throws SQLException {
