+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/test/java/org/apache/shardingsphere/proxy/backend/response/header/query/QueryHeaderBuilderFactoryTest.java
@@ -28,13 +28,13 @@ import static org.junit.Assert.assertTrue;
 public final class QueryHeaderBuilderFactoryTest {
     
     @Test
-    public void assertImplementedDatabaseType() {
+    public void assertNewInstance() {
         QueryHeaderBuilder actual = QueryHeaderBuilderFactory.newInstance(new PostgreSQLDatabaseType());
         assertTrue(actual instanceof PostgreSQLQueryHeaderBuilder);
     }
     
     @Test
-    public void assertUnsupportedDatabaseType() {
+    public void assertNewInstanceWithUnsupportedDatabaseType() {
         QueryHeaderBuilder actual = QueryHeaderBuilderFactory.newInstance(new OracleDatabaseType());
         assertTrue(actual instanceof MySQLQueryHeaderBuilder);
     }
